requirements.txt
fastapi
uvicorn
botocore
boto3
python-multipart
python-dotenv
h11
pip
wheel
PyYAML
anyio
sniffio
s3transfer
jmespath
setuptools
click
starlette
pydantic
websockets
passlib
six
httptools
watchfiles
python-dateutil
urllib3
jose